test(hub): consolidate delete-marker predicate tests into record-keys/tombstone.test.ts (#589)
