+++ packages/hub/__tests__/record-keys/tombstone.test.ts
@@ -5,9 +5,12 @@
  * (#304): {@link buildTombstone} mints the residue shape, {@link isTombstone}
  * recognises it on the read path. End-to-end shred behaviour (history, ledger,
  * subject index) is covered by forget.test.ts — here we pin the primitives.
+ *
+ * Also covers delete markers (#589) which are a distinct envelope predicate.
  */
 import { describe, it, expect } from 'vitest'
 import { isTombstone, buildTombstone } from '../../src/kernel/enclave/index.js'
+import { isDeleteMarker, buildDeleteMarker, isTombstoneShape } from '../../src/kernel/enclave/record-keys/tombstone.js'
 import { NOYDB_FORMAT_VERSION, type EncryptedEnvelope } from '../../src/kernel/types.js'
 
 const live: EncryptedEnvelope = {
@@ -60,5 +63,28 @@ describe('isTombstone', () => {
   it('is always false on an unencrypted collection', () => {
     const { _cek, ...liveNoCek2 } = { ...live, _iv: '', _data: '' }
     expect(isTombstone(liveNoCek2, false)).toBe(false)
+  })
+})
+
+describe('delete marker predicate (#589)', () => {
+  it('isDeleteMarker recognises _del:true and nothing else', () => {
+    expect(isDeleteMarker({ _noydb: 1, _v: 6, _ts: 'x', _iv: '', _data: '', _del: true })).toBe(true)
+    expect(isDeleteMarker({ _noydb: 1, _v: 1, _ts: 'x', _iv: 'iv', _data: 'ct' })).toBe(false)
+    expect(isDeleteMarker({ _noydb: 1, _v: 1, _ts: 'x', _iv: '', _data: '' })).toBe(false) // forget tombstone
+  })
+  it('a delete marker is NOT a forget tombstone (predicates never overlap)', () => {
+    const marker = buildDeleteMarker(6, 'alice')
+    expect(isDeleteMarker(marker)).toBe(true)
+    expect(isTombstoneShape(marker)).toBe(false)             // guarded by _del !== true
+    expect(isTombstone(marker, true)).toBe(false)
+  })
+  it('buildDeleteMarker mints the marker shape at the given version', () => {
+    const m = buildDeleteMarker(6, 'alice')
+    expect(m).toMatchObject({ _noydb: 1, _v: 6, _iv: '', _data: '', _del: true, _by: 'alice' })
+    expect(typeof m._ts).toBe('string')
+    expect(m._cek).toBeUndefined()
+  })
+  it('a forget tombstone is still a tombstone (unchanged)', () => {
+    expect(isTombstoneShape({ _noydb: 1, _v: 3, _ts: 'x', _iv: '', _data: '' })).toBe(true)
   })
 })
